Final1

+++ routes/web.php
@@ -7,6 +7,8 @@ use App\Http\Controllers\ProductosController;
 use App\Http\Controllers\UsersController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\EmpleadoController;
+use App\Models\User;
+
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -19,32 +21,35 @@ use App\Http\Controllers\EmpleadoController;
 */
 
 Route::get('/', function () { return view('index'); });
-Route::get('/login/gerente',[IngresoController::class, 'index'])->name('loginGerente');
+Route::get('/login/gerente',[IngresoController::class, 'index'])->name('login');
 Route::post('/login/gerente',[IngresoController::class,'validacionAdmin'])->name('login.store');
 
 
-Route::get('/soporte/gerente',[UsersController::class, 'indexsoporte'])->name('soporteGerente');
-Route::get('/menu/gerente',[UsersController::class, 'indexmenu'])->name('menuGerente');
-Route::get('/personal/gerente',[UsersController::class, 'indexpersonal'])->name('personalGerente');
-Route::get('/pedidos/gerente',[UsersController::class, 'indexpedidos'])->name('pedidosGerente');
-Route::get('/mesas/gerente',[UsersController::class, 'indexmesas'])->name('mesasGerente');
+Route::get('/soporte/gerente',[UsersController::class, 'indexsoporte'])->name('soporteGerente')->middleware('auth');
+Route::get('/menu/gerente',[UsersController::class, 'indexmenu'])->name('menuGerente')->middleware('auth');
+Route::get('/personal/gerente',[UsersController::class, 'indexpersonal'])->name('personalGerente')->middleware('auth');
+Route::get('/pedidos/gerente',[UsersController::class, 'indexpedidos'])->name('pedidosGerente')->middleware('auth');
+Route::get('/mesas/gerente',[UsersController::class, 'indexmesas'])->name('mesasGerente')->middleware('auth');
 
-Route::get('agregar/personal/gerente',[UsersController::class, 'indexagregarpersonal'])->name('agregarPersonalGerente');
+Route::get('agregar/personal/gerente',[UsersController::class, 'indexagregarpersonal'])->name('agregarPersonalGerente')->middleware('auth');
 
-Route::get('/agregar/mesa/gerente',[UsersController::class, 'indexagregarmesa'])->name('agregarMesaGerente');
+Route::get('/agregar/mesa/gerente',[UsersController::class, 'indexagregarmesa'])->name('agregarMesaGerente')->middleware('auth');
 Route::get('agregar/mesa/gerente',[UsersController::class, 'indexagregarmesa'])->name('agregarMesaGerente');
 
 //Vistas de Menu Gerente
-Route::get('comidas/gerente',[ProductosController::class,'comidas'])->name('comidasGerente');
-Route::post('comidas/store', [ProductosController::class,'store'])->name('storeGerente');
-Route::get('comida/edit/{id}', [ProductosController::class,'edit'])->name('editGerente');
-Route::put('comidas/update/{id}',[ProductosController::class, 'update'])->name('updateGerente');
-Route::delete('comida/delete{id}',[ProductosController::class, 'destroy'])->name('destroyGerente');
+Route::get('comidas/gerente',[ProductosController::class,'comidas'])->name('comidasGerente')->middleware('auth');
+Route::post('comidas/store', [ProductosController::class,'store'])->name('storeGerente')->middleware('auth');
+Route::get('comida/edit/{id}', [ProductosController::class,'edit'])->name('editGerente')->middleware('auth');
+Route::put('comidas/update/{id}',[ProductosController::class, 'update'])->name('updateGerente')->middleware('auth');
+Route::delete('comida/delete{id}',[ProductosController::class, 'destroy'])->name('destroyGerente')->middleware('auth');
 
 
 
-Route::get('bebidas/gerente',[ProductosController::class, 'bebidas'])->name('bebidasGerente');
-Route::get('postres/gerente',[ProductosController::class, 'postres'])->name('postresGerente');
+Route::get('bebidas/gerente',[ProductosController::class, 'bebidas'])->name('bebidasGerente')->middleware('auth');
+Route::get('postres/gerente',[ProductosController::class, 'postres'])->name('postresGerente')->middleware('auth');
+
+Route::get('/destroy',[UsersController::class,'destroy'])->name('destroy')->middleware('auth');
+
 
 //CRUD MENU GERENTE
 
@@ -62,10 +67,10 @@ Route::get('notificacionEnvio/cliente',[ClienteController::class, 'notificacionE
 // Route::get('pedidos/cliente',[ClienteController::class, 'pedidos'])->name('pedidosCliente');
 Route::get('descripcionProducto/cliente',[ClienteController::class, 'descripcionProducto'])->name('descripcionProductoCliente');
 
-
-
-
-Route::get('pp',[ClienteController::class, 'pp'])->name('pp');
-
 //Ruta para los meseros(empledos)
 Route::resource('empleado',EmpleadoController::class);
+
+
+//Prubas de interfaz
+Route::get('pp',[ClienteController::class, 'pp'])->name('pp');
+
